feat: Refactor Json class and add Validator type
This commit introduces a new type, Validator, which is used to validate
the data read from the JSON file. The Json class constructor and _read
method have been refactored to use this new type. This change simplifies
the code and makes it more robust by throwing an error when invalid data
is encountered.