feat: Add TOML support to Limo
This commit introduces support for TOML files in the Limo library. A new
class, `Toml`, has been added which allows reading and writing of TOML
files. This class also supports validation of the data read from the TOML
files. Unit tests have been added to ensure the correct functionality of
the new feature.