feat: Add YAML support to Limo
This commit introduces YAML support to the Limo library. A new class
'Yaml' has been added which allows reading and writing of YAML files.
The Yaml class also supports validation of data similar to other classes
in the library. Corresponding unit tests have been added to ensure the
correctness of the new functionality.