feat: Replace jsr workflow with release workflow
The jsr.yaml workflow has been deleted and replaced with a new
release.yaml workflow. The new workflow maintains the same functionality
as the old one, but it also includes a new release job that runs after
the jsr job. This release job uses the npm:changelogithub command to
generate a changelog.